removed the forced type converstion

+++ test/test_ops.py
@@ -1987,8 +1987,7 @@ class TestOps(unittest.TestCase):
                        lambda data, target: torch_nll_loss(torch_log_softmax(data, dim = 1), target,
                                                            weight = weight_torch, reduction=reduction),
                        lambda data, target: Tensor(data.nll_loss(target, weight = weight_t, reduction=reduction).numpy()),
-                       forward_only=True, vals = [np.random.randn(3, 5).astype(np.float64), np.array([1, 0, 4])])
-
+                       forward_only=True, vals = [np.random.randn(3, 5), np.array([1, 0, 4])])
 
   def test_cross_entropy(self):
     reduction_types = ['none', 'sum', 'mean']
@@ -2004,9 +2003,8 @@ class TestOps(unittest.TestCase):
         helper_test_op(None,
                        lambda data, target: torch_cross_entropy(data, target,
                                                            weight = weight_torch, reduction=reduction),
-                       lambda data, target: Tensor(data.cross_entropy(target, weight = weight_t, reduction=reduction).numpy(), dtype=dtypes.float64),
+                       lambda data, target: Tensor(data.cross_entropy(target, weight = weight_t, reduction=reduction).numpy()),
                        forward_only=True, vals = [np.random.randn(3, 5), [1, 0, 4]])
-
 
   def test_masked_fill(self):
     helper_test_op([(32,10)], lambda x: x.masked_fill((x>0.1).detach(), -math.inf))
